requirements.txt
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3